BCD_Models/ml_service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
tensorflow==2.19.0
scikit-learn==1.6.1
numpy==1.26.4
pandas==2.2.3
joblib==1.4.2
Pillow==10.4.0
python-multipart==0.0.9
